feat(ui): move Queue toggle to a floating tab, out of the transport bar (KAMP-612)
